replace pbf and xml parser with libosmium
In this initial step libosmium only functions as another parser
frontend, replacing the pbf and xml parsers. o5m is still parsed
with the native parser.

Other notable changes include:

- refactor of parse.hpp
- retire UTFsanitizer and remove sanitize option
- make C++11 compiler mandatory (required by libosmium)
- update pbf proto descriptions

+++ osmtypes.hpp
@@ -59,11 +59,11 @@ public:
 
   tag *find(const std::string &key) {  return const_cast<tag *>(_find(key)); }
 
-  size_t indexof(const std::string &key) const
+  int indexof(const std::string &key) const
   {
       for (size_t i = 0; i < size(); ++i)
           if (at(i).key == key)
-              return i;
+              return int(i);
 
       return -1;
   }
